Added getheight & getweight, cleaned up comments and streamlined looping
I put the entire program in a pointless loop. I needed to do that because whenever I need to loop through all the pokemon on the website I needed to manually add the for loop line, indent everything, and then un-indent everything when I finished. To save time I'll just leave everything in a 1 iteration for loop while working on the code.